fix bug

+++ basic_nn_MNIST/main.py
@@ -71,8 +71,8 @@ def loss_fn(model, features, config, training):
     model_output = model(inputs, training=training)
     
     classification_loss = tf.losses.binary_crossentropy(
-        tf.one_hot(label, axis=-1, depth=config.num_classes),
-        model_output, from_logits=False, dtype=tf.int16)
+        tf.one_hot(tf.cast(label, dtype=tf.int16), axis=-1, depth=config.num_classes),
+        model_output, from_logits=False)
     mean_classification_loss = tf.reduce_mean(classification_loss) 
 
     accuracy = tf.reduce_mean(
